+++ packages/label-editor/src/index.ts
@@ -15,4 +15,3 @@ export { default as LaPostePanel } from './lib/components/LaPostePanel.svelte'; 
 export {default as BatchPanel}from'./lib/components/BatchPanel.svelte';export{default as DirectPrintPanel}from'./lib/components/DirectPrintPanel.svelte';export{default as LocalServicePanel}from'./lib/components/LocalServicePanel.svelte';export{default as MediaPanel}from'./lib/components/MediaPanel.svelte';export{default as AssetPanel}from'./lib/components/AssetPanel.svelte';
 export{default as JobRecoveryPanel}from'./lib/components/JobRecoveryPanel.svelte';
 export{default as LibraryPanel}from'./lib/components/LibraryPanel.svelte';
-import './style.css';